fix: debugger failing on Node <=12
Fixes #1624

Updates @vscode/l10n to allow it to be tree-shaken away.

Also moves `checkContentHash` that likewise had a dependency on more
modern language features. I think webpack was just extra aggressive
about tree shaking before and assumed `new Hasher()` was side effect
free. Which is was, it's just a big assumption to have made.

And when I had updated versions in our pipelines, I accidentally updated
the minspec to no longer target Node 8 🤦‍♂️. Targeting the latest 10 should
be good. Node 8 thankfully <0.25% of users nowadays.
